removed Items as singleton
We discovered that const variables are static viariables as well and can be accessed from static functions. This offered the opportunity to make the Itmes class an Object instead of Node and removed as singleton. ItemCreator was removed in this process as well and his functionality moved to the Items class.